// src/lib/api/generated/model/sourceChunk.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Myelin Base RAG API
 * API documentation for Myelin Base RAG backend
 * OpenAPI spec version: 1.0.0
 */
import type { SourceChunkType } from "./sourceChunkType";
import type { SourceChunkSource } from "./sourceChunkSource";

export interface SourceChunk {
  type: SourceChunkType;
  source: SourceChunkSource;
}